Fixed: 스케줄 추가, 삭제시 Confirm 추가

+++ albalog-client/src/components/Modal/ScheduleInfoModal.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import client from 'utils/api';
 import './ScheduleInfoModal.scss';
 
@@ -6,16 +6,17 @@ const ScheduleInfoModal = ({ handleInfoModal, employee, locationId }) => {
   const handleDeleteSubmit = async (e) => {
     e.preventDefault();
 
-    try {
-      const response = await client.delete(
-        `shift/${employee.index}/location/${locationId}/employee/${employee.staffId}/delete`,
-      );
-      handleInfoModal();
-      if (response.status === 200) {
-        window.location.replace(`/admin/${locationId}/schedule`); // 페이지 이동 후 새로고침
+    const isConfirm = window.confirm('정말 삭제하시겠습니까?');
+
+    if (isConfirm) {
+      try {
+        await client.delete(
+          `shift/${employee.index}/location/${locationId}/employee/${employee.staffId}/delete`,
+        );
+        handleInfoModal();
+      } catch (e) {
+        console.error(e);
       }
-    } catch (e) {
-      console.error(e);
     }
   };
 
@@ -23,17 +24,20 @@ const ScheduleInfoModal = ({ handleInfoModal, employee, locationId }) => {
     <div className="info-modal-container">
       <div className="info-modal-wrap">
         <h2 className="title">스케줄 정보</h2>
-        <div className="employee-detail">
-          <p className="content">🤔 이름: {employee.title}</p>
-          <p className="content">
-            {`📅 날짜: ${employee.start.getFullYear()}년 ${
-              employee.start.getMonth() + 1
-            }월 ${employee.start.getDate()}일`}
-          </p>
-          <p className="content">
-            {`🕑 근무시간: ${employee.start.getHours()}시 ${employee.start.getMinutes()}분 ~ ${employee.end.getHours()}시 ${employee.end.getMinutes()}분`}
-          </p>
-        </div>
+        {employee && (
+          <div className="employee-detail">
+            <p className="content">🤔 이름: {employee.title}</p>
+            <p className="content">
+              {`📅 날짜: ${employee.start.getFullYear()}년 ${
+                employee.start.getMonth() + 1
+              }월 ${employee.start.getDate()}일`}
+            </p>
+            <p className="content">
+              {`🕑 근무시간: ${employee.start.getHours()}시 ${employee.start.getMinutes()}분 ~ ${employee.end.getHours()}시 ${employee.end.getMinutes()}분`}
+            </p>
+          </div>
+        )}
+
         <button className="btn-del" onClick={handleDeleteSubmit}>
           삭제하기
         </button>
